requirements.txt
eyed3~=0.9.8
pillow~=11.3.0
mutagen~=1.47.0